Consider step-functions when determining export
We will not export the full demand of a cost-function producer if it is
semi-competitive with the local country. For example, if the local energy
cost is 50.0, and the foreign country has a producer which looks like:

   1 plant  - 45.0   # <- competitive
   2 plants - 50.0   # <- competitive
   3 plants - 55.0   # <- uncompetitive

... then we only export one plant-worth of electricity, since exceeding
that would be more expensive to the foreign country than generating it
themselves.

Ref: quintel/merit#115
See: https://github.com/quintel/merit/issues/115#issuecomment-50442351